requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
fastapi==0.115.8
frozenlist==1.5.0
gidgethub==5.3.0
h11==0.14.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uritemplate==4.1.1
uvicorn==0.34.0
yarl==1.18.3
